- Added Game Loaded method to insure mod is loaded.
