+++ BetterCemeteryAI/AIHelper.cs
@@ -1,20 +1,77 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 
 namespace BetterCemeteryAI
 {
     public static class AIHelper
     {
-        public static void ChangeBuildingAI(Building b, Type AIType)
+
+        public static void ApplyNewAIToBuilding(Building b)
         {
+            try
+            {
+                if (b.Info.m_class.name.Equals("DeathCare Facility"))
+                {
+                    ChangeBuildingAI(b, typeof(NewCemetaryAI));
+                    return;
+                }
+            }
+            catch (Exception e)
+            {
+                LogHelper.Error(e.ToString());
+            }
+        }
+
+        private static void ChangeBuildingAI(Building b, Type AIType)
+        {
+            //Delete old AI
             var oldAI = b.Info.gameObject.GetComponent<PrefabAI>();
             UnityEngine.Object.DestroyImmediate(oldAI);
-            var newAIInfo = new PrefabAIInfo(AIType);
-            var newAI = (PrefabAI)b.Info.gameObject.AddComponent(newAIInfo.type);
-            PrefabAIInfo.TryCopyAttributes(oldAI, newAI, false);
+
+            //Add new AI
+            var newAI = (PrefabAI)b.Info.gameObject.AddComponent(AIType);
+            TryCopyAttributes(oldAI, newAI, false);
             b.Info.InitializePrefab();
+        }
+
+        private static void TryCopyAttributes(PrefabAI src, PrefabAI dst, bool safe = true)
+        {
+            var oldAIFields = src.GetType()
+                .GetFields(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.FlattenHierarchy);
+            var newAIFields = dst.GetType()
+                .GetFields(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.FlattenHierarchy);
+
+            var newAIFieldDic = new Dictionary<string, FieldInfo>(newAIFields.Length);
+            foreach (var field in newAIFields)
+            {
+                newAIFieldDic.Add(field.Name, field);
+            }
+
+            foreach (var fieldInfo in oldAIFields)
+            {
+                bool copyField = !fieldInfo.IsDefined(typeof(NonSerializedAttribute), true);
+
+                if (safe && !fieldInfo.IsDefined(typeof(CustomizablePropertyAttribute), true)) copyField = false;
+
+                if (copyField)
+                {
+                    FieldInfo newAIField;
+                    newAIFieldDic.TryGetValue(fieldInfo.Name, out newAIField);
+                    try
+                    {
+                        if (newAIField != null && newAIField.GetType().Equals(fieldInfo.GetType()))
+                        {
+                            newAIField.SetValue(dst, fieldInfo.GetValue(src));
+                        }
+                    }
+                    catch (NullReferenceException)
+                    {
+                    }
+                }
+            }
         }
     }
 }
